[ui] Set Y axis to 100 if unit is percentage #136
Before this fix, the RAM/CPU/disk usage charts may look
like they're showing anomalous behavior, even if the chart
is flat at 40%, because plotly will resize the canvas
to fit the chart.

By setting the Y axis to 100 when percentage is used
we make the chart more readable.

Implements and closes #136