[change] ConfigApplied triggered on config_status_changed instead of config_modified
It does not make sense to trigger this check when the config is modified,
because the result will be always that config.status == 'modified'.
Moreover, due to the recent changes in openwisp-controller,
this action can now be executed also when a config has been just created
and m2m templates are added to it.

Rather than changing the code to check whether the signal is being
sent from m2m changes or not, it is a lot more useful to just
switch to the config_status_changed signal, which will trigger
the check also when the configuration is flagged as applied by
the system, making the check a lot more proactive.