[fix] Fixed wifi session tabular inline error #504
If wifi sessions are created in the background
while the user is still on the device change page,
saving the device object can raise a formset error.
To prevent this, we can check if wifi sessions exist
before saving the device object and provide default
values for the complaining formset fields.

Fixes #504