[chores] Do not allow selecting columns when exporting devices
The django-export-import dependency was upgraded in
openwisp-controller which now allows selecting the columns during
export operation.

Since this can cause issues during import of files with limited
fields, we decided to rollback to the old implementation which
exports all the fields of the device model.

Related https://github.com/openwisp/openwisp-controller/pull/915